Renamer chops characters and dry runs
- It's also set up to work properly with its dependencies

+++ bin/renamer.js
@@ -12,38 +12,56 @@ try {
   var minimist = require('minimist');
 } catch (err) {
   console.error('Failed to load minimist module.');
-  console.error('  Run \'npm install\' in the ~/bin directory.');
+  var dir = path.dirname(__filename);
+  console.error('  Run \'npm install\' in the ' + dir + ' directory.');
   process.exit(1);
 }
 
 function rename(filename, options) {
+  // Separate path
   var ext = path.extname(filename);
   var base = path.basename(filename, ext);
 
+  // Lowercase
   if (options.hasOwnProperty('l') || options.hasOwnProperty('lowercase')) {
     ext = ext.toLowerCase();
     base = base.toLowerCase();
   }
 
+  // Chop
+  if (options.hasOwnProperty('c') && typeof options.c == 'number') {
+    base = base.substring(options.c);
+  }
+  if (options.hasOwnProperty('chop') && typeof options.chop == 'number') {
+    base = base.substring(options.chop);
+  }
+
+  // Prefix
   if (options.hasOwnProperty('p') && typeof options.p == 'string') {
     base = options.p + base;
   }
-
   if (options.hasOwnProperty('prefix') && typeof options.prefix == 'string') {
     base = options.prefix + base;
   }
 
+  // Today's date
   if (options.hasOwnProperty('t') || options.hasOwnProperty('prefix-today')) {
     var today = new Date();
     base = today.toJSON().substr(0, 10) + '_' + base;
   }
 
+  // Underscore
   if (options.hasOwnProperty('u') || options.hasOwnProperty('underscore')) {
     base = base.replace(/ /g, '_');
   }
 
+  // Combine new name
   var newname = path.join(path.dirname(filename), base + ext);
-  fs.renameSync(filename, newname);
+  // If not dry-running, do the rename
+  if (!options.hasOwnProperty('d') && !options.hasOwnProperty('dry-run')) {
+    fs.renameSync(filename, newname);
+  }
+  // Print change
   console.log(filename + ' -> ' + newname);
 }
 
@@ -69,7 +87,9 @@ function analyzeList(args) {
 }
 
 function usage() {
-  console.log('Usage: ' + process.argv[1] + ' -[lptu] -- {files or directories}');
+  console.log('Usage: ' + process.argv[1] + ' -[cdlptu] -- {files or directories}');
+  console.log('  -c or --chop {number} - chop characters off the front');
+  console.log('  -d or --dry-run - dry run, don\'t actually rename files');
   console.log('  -l or --lowercase - change captials to lowercase');
   console.log('  -p or --prefix {prefix} - prefix with arbitrary information');
   console.log('  -t or --prefix-today - prefix with today\'s YYYY-MM-DD date');
